fix: review recommendations - part 3
