+++ backend/api/src/main/kotlin/io/tolgee/controllers/PublicController.kt
@@ -3,33 +3,26 @@ package io.tolgee.controllers
 import com.fasterxml.jackson.databind.node.TextNode
 import io.swagger.v3.oas.annotations.Operation
 import io.tolgee.configuration.tolgee.AuthenticationProperties
-import io.tolgee.configuration.tolgee.TolgeeProperties
 import io.tolgee.constants.Message
 import io.tolgee.dtos.request.auth.SignUpDto
 import io.tolgee.dtos.security.LoginRequest
-import io.tolgee.dtos.sso.SsoTenantConfig
 import io.tolgee.exceptions.AuthenticationException
 import io.tolgee.exceptions.BadRequestException
 import io.tolgee.exceptions.DisabledFunctionalityException
-import io.tolgee.exceptions.NotFoundException
 import io.tolgee.hateoas.invitation.PublicInvitationModel
 import io.tolgee.hateoas.invitation.PublicInvitationModelAssembler
-import io.tolgee.model.enums.ThirdPartyAuthType
 import io.tolgee.openApiDocs.OpenApiHideFromPublicDocs
 import io.tolgee.security.authentication.JwtService
 import io.tolgee.security.payload.JwtAuthenticationResponse
 import io.tolgee.security.ratelimit.RateLimited
-import io.tolgee.security.service.thirdParty.ThirdPartyAuthDelegate
-import io.tolgee.security.thirdParty.ThirdPartyUserHandler
-import io.tolgee.security.thirdParty.data.ThirdPartyUserDetails
 import io.tolgee.service.EmailVerificationService
-import io.tolgee.service.TenantService
 import io.tolgee.service.invitation.InvitationService
 import io.tolgee.service.security.MfaService
 import io.tolgee.service.security.ReCaptchaValidationService
 import io.tolgee.service.security.SignUpService
 import io.tolgee.service.security.UserAccountService
 import io.tolgee.service.security.UserCredentialsService
+import io.tolgee.service.security.thirdParty.ThirdPartyAuthenticationService
 import jakarta.validation.Valid
 import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.NotNull
@@ -48,7 +41,6 @@ import org.springframework.web.bind.annotation.RestController
 @AuthenticationTag
 class PublicController(
   private val jwtService: JwtService,
-  private val properties: TolgeeProperties,
   private val userAccountService: UserAccountService,
   private val emailVerificationService: EmailVerificationService,
   private val reCaptchaValidationService: ReCaptchaValidationService,
@@ -56,9 +48,7 @@ class PublicController(
   private val mfaService: MfaService,
   private val userCredentialsService: UserCredentialsService,
   private val authProperties: AuthenticationProperties,
-  private val thirdPartyAuthDelegates: List<ThirdPartyAuthDelegate>,
-  private val thirdPartyUserHandler: ThirdPartyUserHandler,
-  private val tenantService: TenantService,
+  private val thirdPartyAuthenticationService: ThirdPartyAuthenticationService,
   private val publicInvitationModelAssembler: PublicInvitationModelAssembler,
   private val invitationService: InvitationService,
 ) {
@@ -137,22 +127,13 @@ class PublicController(
     @RequestParam(value = "invitationCode", required = false) invitationCode: String?,
     @RequestParam(value = "domain", required = false) domain: String?,
   ): JwtAuthenticationResponse {
-    val delegate =
-      thirdPartyAuthDelegates.find { it.name == serviceType }
-        ?: throw NotFoundException(Message.SERVICE_NOT_FOUND)
-
-    if (properties.internal.fakeThirdPartyLogin && code?.startsWith(DUMMY_CODE_PREFIX) == true) {
-      val username = code.removePrefix(DUMMY_CODE_PREFIX)
-      val data =
-        DummyThirdPartyUserDetails(
-          username = username,
-          name = username,
-          thirdPartyAuthType = delegate.preferredAuthType,
-        )
-      return fakeThirdPartyLogin(data, invitationCode, domain)
-    }
-
-    return delegate.getTokenResponse(code, invitationCode, redirectUri, domain)
+    return thirdPartyAuthenticationService.authenticate(
+      serviceType = serviceType,
+      code = code,
+      redirectUri = redirectUri,
+      invitationCode = invitationCode,
+      domain = domain,
+    )
   }
 
   @GetMapping("/invitation_info/{code}")
@@ -162,44 +143,5 @@ class PublicController(
   ): PublicInvitationModel {
     val invitation = invitationService.getInvitation(code)
     return publicInvitationModelAssembler.toModel(invitation)
-  }
-
-  private fun fakeThirdPartyLogin(
-    data: DummyThirdPartyUserDetails,
-    invitationCode: String?,
-    domain: String?,
-  ): JwtAuthenticationResponse {
-    var tenant: SsoTenantConfig? = null
-    if (data.thirdPartyAuthType == ThirdPartyAuthType.SSO) {
-      tenant = tenantService.getEnabledConfigByDomain(domain)
-      if (tenant.global) {
-        // We have found global tenant - fix the auth type accordingly
-        data.thirdPartyAuthType = ThirdPartyAuthType.SSO_GLOBAL
-      }
-    }
-    val user =
-      thirdPartyUserHandler.findOrCreateUser(
-        ThirdPartyUserDetails(
-          authId = "dummy_auth_id",
-          username = data.username,
-          name = data.name,
-          thirdPartyAuthType = data.thirdPartyAuthType,
-          invitationCode = invitationCode,
-          refreshToken = null,
-          tenant = tenant,
-        ),
-      )
-    val jwt = jwtService.emitToken(user.id)
-    return JwtAuthenticationResponse(jwt)
-  }
-
-  data class DummyThirdPartyUserDetails(
-    var username: String,
-    var name: String,
-    var thirdPartyAuthType: ThirdPartyAuthType,
-  )
-
-  companion object {
-    const val DUMMY_CODE_PREFIX = "this_is_dummy_code_"
   }
 }
